requirements.txt
Django>=4.2
dnspython
dkimpy
requests
beautifulsoup4
lxml
watchdog
weasyprint
